+++ src/components/matches-table/matches-table.component.js
@@ -27,7 +27,7 @@ module.exports = {
             columns: columns
         }
     },
-    onShowModal(matchId) {
-        this.emit('show-modal', matchId);
+    onShowModal(matchId, modal) {
+        this.emit('show-modal', matchId, modal);
     }
 }