+++ src/components/spectate/components/spectate-header/spectate-header.component.js
@@ -1,13 +1,21 @@
+const localStorage = require('../../../../util/localstorage');
+
 module.exports = {
-    onInput(input) {
+    onCreate(input) {
         this.state = {
             round: input.round,
             enableAnnouncement: false
         }
     },
+    onMount() {
+        this.state.enableAnnouncement = localStorage.getBool('spectate-announcements', false);
+        this.emit('enable-announcement', this.state.enableAnnouncement);
+    },
     enableVoiceAnnouncement(event) {
-        var enable = !this.state.enableAnnouncement;
+        const enable = !this.state.enableAnnouncement;
         this.emit('enable-announcement', enable);
         this.state.enableAnnouncement = enable;
+
+        localStorage.set('spectate-announcements', enable);
     }
 };